requirements.txt
beautifulsoup4==4.12.2
langchain==0.0.277
playwright==1.37.0
python-dotenv==1.0.0
pyunsplash==1.0.0rc2
Requests==2.31.0
colorama==0.4.6
fake-useragent==1.2.1
pyperclip==1.8.2
selenium==4.11.2
tempmail-python==2.3.0
selenium
undetected_chromedriver
fake_useragent
